+++ semseg/dataloader/cityscapes_loader.py
@@ -256,7 +256,7 @@ class cityscapesLoader(data.Dataset):
             g[temp == l] = self.label_colours[l][1]
             b[temp == l] = self.label_colours[l][2]
 
-        rgb = np.zeros((temp.shape[0], temp.shape[1], 3))
+        rgb = np.zeros((temp.shape[0], temp.shape[1], 3), dtype=np.float32)
         rgb[:, :, 0] = r / 255.0
         rgb[:, :, 1] = g / 255.0
         rgb[:, :, 2] = b / 255.0
